requirements.txt
altgraph==0.17
future==0.18.2
keyboard==0.13.5
MouseInfo==0.1.3
numpy==1.20.1
opencv-python==4.5.1.48
pefile==2019.4.18
Pillow==8.1.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pywin32-ctypes==0.2.0
pyzbar==0.1.8
